requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==25.1.0; python_version >= '3.8'
django==5.1.7; python_version >= '3.10'
django-filter==25.1; python_version >= '3.9'
djangorestframework==3.15.2; python_version >= '3.8'
drf-spectacular==0.28.0; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
markdown==3.7; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.23.1; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
uritemplate==4.1.1; python_version >= '3.6'
